refactor: split component into smaller components
